+++ src/components/Buttons/BackButton/backButton.module.scss
@@ -1,0 +1,3 @@
+.backButton {
+  width: 20px;
+}
