+++ src/components/Buttons/CompleteButton/completeButton.module.scss
@@ -1,0 +1,3 @@
+.completeButton {
+  font-size: 16px;
+}
