+++ src/components/Buttons/CompleteButton/index.jsx
@@ -1,0 +1,7 @@
+import styles from './completeButton.module.scss'
+
+const completeButton = () => {
+  return <button className={styles.completeButton}>완료</button>
+}
+
+export default completeButton
