+++ src/components/Gnb/gnb.module.scss
@@ -1,0 +1,7 @@
+.nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding: 24px 20px;
+}
