+++ src/routes/MyChallengePage/myChallengePage.module.scss
@@ -1,12 +1,6 @@
 .myChallengePage {
   width: 100%;
 
-  & > nav {
-    width: 100%;
-    padding: 24px 0;
-    padding-left: 20px;
-  }
-
   .content {
     display: flex;
     flex-direction: column;
