+++ src/routes/SignupPage/signupPage.module.scss
@@ -42,11 +42,12 @@
     justify-content: space-between;
     width: 288px;
     margin-bottom: 65px;
-    &>h1{
-      font-family: 'Work Sans';
+
+    & > h1 {
+      font-family: "Work Sans";
+      font-size: 10px;
       font-style: normal;
       font-weight: 400;
-      font-size: 10px;
       line-height: 37px;
       color: aqua;
     }
